requirements.txt
Flask==3.0.3
flask-cors==6.0.2
PyJWT==2.8.0
bcrypt==4.1.2
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
requests==2.32.3
aiohttp==3.10.5
tenacity==8.2.3
pytest==7.4.4
werkzeug==3.0.3
